(BSR)[API] refactor: Move educational exceptions to correct module
