--- api/src/pcapi/connectors/api_adage.py
@@ -1,11 +1,0 @@
-# FIXME: move these exceptions to `core.offerers.exceptions`?
-class AdageException(Exception):
-    def __init__(self, message, status_code, response_text):  # type: ignore [no-untyped-def]
-        self.message = message
-        self.status_code = status_code
-        self.response_text = response_text
-        super().__init__(message)
-
-
-class CulturalPartnerNotFoundException(Exception):
-    pass
