+++ api/src/pcapi/core/educational/adage_backends/logger.py
@@ -1,8 +1,8 @@
 import logging
 
-from pcapi.connectors.api_adage import CulturalPartnerNotFoundException
 from pcapi.connectors.serialization.api_adage_serializers import AdageVenue
 from pcapi.core.educational.adage_backends.base import AdageClient
+from pcapi.core.educational.exceptions import CulturalPartnerNotFoundException
 from pcapi.core.educational.models import AdageApiResult
 from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingEdition
 from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingResponse
